refactor: Position 생성자 관련 이슈
Position 생성자를 하나로 확정 + 패키지 액세스
Positions 패키지 이동
ALL 패키지 이동에 따른 수정
